+++ .umirc.ts
@@ -18,4 +18,5 @@ export default defineConfig({
   publicPath: '/react-admin-kit/',
   base: '/react-admin-kit/',
   mfsu: {},
+  dynamicImport: {},
 });