+++ docs/coreComponents/SchemaForm/group/index.tsx
@@ -1,10 +1,11 @@
 import { SchemaForm } from 'react-admin-kit';
+import { Button } from 'antd';
 import { columns } from './columns';
 import { useRef } from 'react';
 
 import type { FormInstance } from 'antd';
 
-const Readonly = () => {
+const Group = () => {
   const formRef = useRef<FormInstance>();
 
   const onFinish = (values: any) => {
@@ -14,19 +15,22 @@ const Readonly = () => {
   return (
     <div>
       <SchemaForm
-        layout="horizontal"
-        grid={true}
-        rowProps={{ gutter: [0, 0] }}
-        colProps={{ span: 8 }}
-        labelCol={{ span: 9 }}
         onFinish={onFinish}
         formRef={formRef}
         columns={columns}
-        readonly={true}
-        initialValues={{ name: '王先生', phone: 1202228888, sex: '男', mail: 'mail.com' }}
+        autoFocusFirstInput={false}
       />
+
+      <div style={{ textAlign: 'end' }}>
+        <Button style={{ marginRight: '10px' }} onClick={() => formRef.current?.resetFields()}>
+          重置
+        </Button>
+        <Button type="primary" onClick={() => formRef.current?.submit()}>
+          提交
+        </Button>
+      </div>
     </div>
   );
 };
 
-export default Readonly;
+export default Group;