+++ src/SettingProvider/context.ts
@@ -1,6 +1,11 @@
 import React from 'react';
-import { ModalFormSetting, ProTableSetting } from './types';
+import { FormUploadProps } from '../FormUpload/types';
+import { ModalFormSetting, ProTableSetting, SettingFormProps } from './types';
 
 export const ProTableContext = React.createContext<ProTableSetting>(undefined);
 
 export const ModalFormContext = React.createContext<ModalFormSetting>(undefined);
+
+export const SchemaFormContext = React.createContext<SettingFormProps | undefined>(undefined);
+
+export const FormUploadContext = React.createContext<FormUploadProps | undefined>(undefined);