+++ src/SettingProvider/index.tsx
@@ -1,12 +1,19 @@
-import { ModalFormContext, ProTableContext } from './context';
+import { FormUploadContext, ModalFormContext, ProTableContext, SchemaFormContext } from './context';
 import { SettingProviderProps } from './types';
 
 const SettingProvider = (props: SettingProviderProps) => {
-  const { proTableSetting, modalFormSetting, children } = props;
+  const { proTableSetting, modalFormSetting, schemaFormSetting, formUploadSetting, children } =
+    props;
 
   return (
     <ProTableContext.Provider value={proTableSetting}>
-      <ModalFormContext.Provider value={modalFormSetting}>{children}</ModalFormContext.Provider>
+      <ModalFormContext.Provider value={modalFormSetting}>
+        <SchemaFormContext.Provider value={schemaFormSetting}>
+          <FormUploadContext.Provider value={formUploadSetting}>
+            {children}
+          </FormUploadContext.Provider>
+        </SchemaFormContext.Provider>
+      </ModalFormContext.Provider>
     </ProTableContext.Provider>
   );
 };