+++ src/SettingProvider/types.ts
@@ -1,5 +1,6 @@
 import { SearchConfig } from '@ant-design/pro-table/es/components/Form/FormRender';
 import type { ModalProps } from 'antd';
+import { FormUploadProps } from '../FormUpload/types';
 import { SchemaFormProps } from '../SchemaForm/types';
 
 export type SettingFormProps = Omit<SchemaFormProps, 'onFinish' | 'columns'>;
@@ -32,5 +33,17 @@ export type SettingProviderProps = {
    * @default           false
    */
   modalFormSetting?: ModalFormSetting;
+
+  /**
+   * @description       SchemaForm组件的设置项
+   * @default           false
+   */
+  schemaFormSetting?: SettingFormProps;
+
+  /**
+   * @description       FormUpload组件的设置项
+   * @default           false
+   */
+  formUploadSetting?: FormUploadProps;
   children: any;
 };