+++ src/index.ts
@@ -9,6 +9,7 @@ export { clearTreeSelectCache } from './BusinessTreeSelectBuilder';
 export { default as SettingProvider } from './SettingProvider';
 export { default as Button } from './Button';
 export { ProForm } from '@ant-design/pro-form';
+export { default as FormUpload } from './FormUpload';
 
 export type { MyProColumnType as ProColumnType, InnerRefType } from './ProTable/types';
 export type { ModalFormInnerRefType } from './ModalForm/types';